Park func_8005ABA0 at 175/-2, and a way to read a u8 parameter off the callee
The lever worth keeping: a `u8` parameter is right not only when the CALL
SITE masks, but when the callee loads its own stack slot with `lbu`. For the
fifth-and-later arguments that is direct evidence, and it is expensive to get
wrong -- declared `s32` the `lw` lands in an argument register and forces a
copy of a real argument out of its own, shifting every register after it.
func_8005ABA0's functions.h prototype was PROVISIONAL and said s32; the
listing says lbu. Fixed, and its one decompiled caller rechecked.

The rest of that function: u32 for the `& 0xFFFF` selector so the compare is
`sltiu`, `(sa * 2) * L` forced through a named intermediate because
`sa * 2 * L` reassociates to `sa * (2 * L)`, and a second name for the third
clamped byte only. Park entry lists the four shapes measured against the
remaining two instructions.

Build green, sha1 verified.

+++ include/functions.h
@@ -417,7 +417,8 @@ void func_80046990(s32 arg0, s32 arg1, s32 arg2);  /* PROVISIONAL */
 void func_8003BF00(void);  /* PROVISIONAL */
 void func_80045F3C(void);  /* PROVISIONAL */
 void func_80046A08(void);  /* PROVISIONAL */
-void func_8005ABA0(u8 *arg0, s32 arg1, s32 arg2, s32 arg3, s32 arg4);  /* PROVISIONAL */
+/* arg4 is a u8: func_8005ABA0 loads its stack slot with `lbu`. */
+void func_8005ABA0(u8 *arg0, s32 arg1, s32 arg2, s32 arg3, u8 arg4);
 void func_80019B2C(u8 *arg0);
 void func_8004544C(void);
